+++ src/main/java/br/com/zalf/prolog/webservice/gente/quiz/relatorios/QuizRelatorioResource.java
@@ -1,8 +1,16 @@
 package br.com.zalf.prolog.webservice.gente.quiz.relatorios;
 
 import br.com.zalf.prolog.webservice.commons.report.Report;
-import br.com.zalf.prolog.webservice.permissao.pilares.Pilares;
+import br.com.zalf.prolog.webservice.commons.util.Optional;
+import br.com.zalf.prolog.webservice.commons.util.Required;
+import br.com.zalf.prolog.webservice.errorhandling.exception.ProLogException;
 import br.com.zalf.prolog.webservice.interceptors.auth.Secured;
+import br.com.zalf.prolog.webservice.interceptors.versioncodebarrier.AppVersionCodeHandler;
+import br.com.zalf.prolog.webservice.interceptors.versioncodebarrier.DefaultAppVersionCodeHandler;
+import br.com.zalf.prolog.webservice.interceptors.versioncodebarrier.VersionCodeHandlerMode;
+import br.com.zalf.prolog.webservice.interceptors.versioncodebarrier.VersionNotPresentAction;
+import br.com.zalf.prolog.webservice.permissao.pilares.Pilares;
+import org.jetbrains.annotations.NotNull;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
@@ -12,86 +20,149 @@ import javax.ws.rs.core.StreamingOutput;
  * Created by Zart on 20/03/17.
  */
 @Path("/quizzes/relatorios")
+@Secured(permissions = Pilares.Gente.Relatorios.QUIZ)
 @Consumes(MediaType.APPLICATION_JSON + ";charset=utf-8")
 @Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
+@AppVersionCodeHandler(
+        implementation = DefaultAppVersionCodeHandler.class,
+        targetVersionCode = 76,
+        versionCodeHandlerMode = VersionCodeHandlerMode.BLOCK_THIS_VERSION_AND_BELOW,
+        actionIfVersionNotPresent = VersionNotPresentAction.BLOCK_ANYWAY)
 public class QuizRelatorioResource {
-
-    private QuizRelatorioService service = new QuizRelatorioService();
+    @NotNull
+    private final QuizRelatorioService service = new QuizRelatorioService();
 
     @GET
-    @Path("/realizados/{codUnidade}/{codModeloQuiz}/csv")
-    @Secured(permissions = Pilares.Gente.Relatorios.QUIZ)
+    @Path("/estratificacao-realizados/csv")
     @Produces("application/csv")
-    public StreamingOutput getEstratificacaoRealizacaoQuizCsv(@PathParam("codUnidade") Long codUnidade,
-                                                              @PathParam("codModeloQuiz") String codModeloQuiz,
-                                                              @QueryParam("dataInicial") long dataInicial,
-                                                              @QueryParam("dataFinal") long dataFinal) {
-        return outputStream -> service.getEstratificacaoRealizacaoQuizCsv(outputStream, codModeloQuiz, codUnidade, dataInicial,
+    public StreamingOutput getEstratificacaoRealizacaoQuizCsv(@QueryParam("codUnidade") @Required Long codUnidade,
+                                                              @QueryParam("codModeloQuiz") @Optional Long codModeloQuiz,
+                                                              @QueryParam("dataInicial") @Required String dataInicial,
+                                                              @QueryParam("dataFinal") @Required String dataFinal) {
+        return outputStream -> service.getEstratificacaoRealizacaoQuizCsv(
+                outputStream,
+                codUnidade,
+                codModeloQuiz,
+                dataInicial,
                 dataFinal);
     }
 
     @GET
-    @Path("/realizados/{codUnidade}/{codModeloQuiz}/report")
-    @Secured(permissions = Pilares.Gente.Relatorios.QUIZ)
-    public Report getEstratificacaoRealizacaoQuizReport(@PathParam("codUnidade") Long codUnidade,
-                                                        @PathParam("codModeloQuiz") String codModeloQuiz,
-                                                        @QueryParam("dataInicial") long dataInicial,
-                                                        @QueryParam("dataFinal") long dataFinal) {
-        return service.getEstratificacaoRealizacaoQuizReport(codModeloQuiz, codUnidade, dataInicial, dataFinal);
+    @Path("/estratificacao-realizados/report")
+    public Report getEstratificacaoRealizacaoQuizReport(
+            @QueryParam("codUnidade") @Required Long codUnidade,
+            @QueryParam("codModeloQuiz") @Optional Long codModeloQuiz,
+            @QueryParam("dataInicial") @Required String dataInicial,
+            @QueryParam("dataFinal") @Required String dataFinal) throws ProLogException {
+        return service.getEstratificacaoRealizacaoQuizReport(
+                codUnidade,
+                codModeloQuiz,
+                dataInicial,
+                dataFinal);
     }
 
     @GET
-    @Path("/cargos/{codUnidade}/{codModeloQuiz}/csv")
-    @Secured(permissions = Pilares.Gente.Relatorios.QUIZ)
+    @Path("/realizacoes-por-cargos/csv")
     @Produces("application/csv")
-    public StreamingOutput getRealizacaoQuizByCargoCsv(@PathParam("codUnidade") Long codUnidade,
-                                                       @PathParam("codModeloQuiz") String codModeloQuiz) {
-        return outputStream -> service.getRealizacaoQuizByCargoCsv(outputStream, codUnidade, codModeloQuiz);
+    public StreamingOutput getRealizacaoQuizByCargoCsv(@QueryParam("codUnidade") @Required Long codUnidade,
+                                                       @QueryParam("codModeloQuiz") @Optional Long codModeloQuiz) {
+        return outputStream -> service.getRealizacaoQuizByCargoCsv(
+                outputStream,
+                codUnidade,
+                codModeloQuiz);
     }
 
     @GET
-    @Path("/cargos/{codUnidade}/{codModeloQuiz}/report")
-    @Secured(permissions = Pilares.Gente.Relatorios.QUIZ)
-    public Report getRealizacaoQuizByCargoReport(@PathParam("codUnidade") Long codUnidade,
-                                                 @PathParam("codModeloQuiz") String codModeloQuiz,
-                                                 @QueryParam("dataInicial") long dataInicial,
-                                                 @QueryParam("dataFinal") long dataFinal) {
-        return service.getRealizacaoQuizByCargoReport(codUnidade, codModeloQuiz);
+    @Path("/realizacoes-por-cargos/report")
+    public Report getRealizacaoQuizByCargoReport(
+            @QueryParam("codUnidade") @Required Long codUnidade,
+            @QueryParam("codModeloQuiz") Long codModeloQuiz) throws ProLogException {
+        return service.getRealizacaoQuizByCargoReport(
+                codUnidade,
+                codModeloQuiz);
     }
 
     @GET
-    @Path("/respostas/{codUnidade}/{codModeloQuiz}/csv")
-    @Secured(permissions = Pilares.Gente.Relatorios.QUIZ)
+    @Path("/estratificacao-respostas/csv")
     @Produces("application/csv")
-    public StreamingOutput getEstratificacaoQuizRespostasCsv(@PathParam("codUnidade") Long codUnidade,
-                                                             @PathParam("codModeloQuiz") String codModeloQuiz) {
-        return outputStream -> service.getEstratificacaoQuizRespostasCsv(outputStream, codUnidade, codModeloQuiz);
+    public StreamingOutput getEstratificacaoQuizRespostasCsv(@QueryParam("codUnidade") @Required Long codUnidade,
+                                                             @QueryParam("codModeloQuiz") @Optional Long codModeloQuiz) {
+        return outputStream -> service.getEstratificacaoQuizRespostasCsv(
+                outputStream,
+                codUnidade,
+                codModeloQuiz);
     }
 
     @GET
-    @Path("/respostas/{codUnidade}/{codModeloQuiz}/report")
-    @Secured(permissions = Pilares.Gente.Relatorios.QUIZ)
-    public Report getEstratificacaoQuizRespostasReport(@PathParam("codUnidade") Long codUnidade,
-                                                       @PathParam("codModeloQuiz") String codModeloQuiz) {
-        return service.getEstratificacaoQuizRespostasReport(codUnidade, codModeloQuiz);
+    @Path("/estratificacao-respostas/report")
+    public Report getEstratificacaoQuizRespostasReport(@QueryParam("codUnidade") @Required Long codUnidade,
+                                                       @QueryParam("codModeloQuiz") @Optional Long codModeloQuiz)
+            throws ProLogException {
+        return service.getEstratificacaoQuizRespostasReport(
+                codUnidade,
+                codModeloQuiz);
     }
 
     @GET
-    @Path("/consolidados/{codUnidade}/csv")
-    @Secured(permissions = Pilares.Gente.Relatorios.QUIZ)
+    @Path("/extrato-geral/csv")
     @Produces("application/csv")
-    public StreamingOutput getExtratoGeralCsv(@PathParam("codUnidade") Long codUnidade,
-                                              @QueryParam("dataInicial") long dataInicial,
-                                              @QueryParam("dataFinal") long dataFinal) {
-        return outputStream -> service.getExtratoGeralCsv(outputStream, codUnidade, dataInicial, dataFinal);
+    public StreamingOutput getExtratoGeralCsv(@QueryParam("codUnidade") @Required Long codUnidade,
+                                              @QueryParam("dataInicial") @Required String dataInicial,
+                                              @QueryParam("dataFinal") @Required String dataFinal) {
+        return outputStream -> service.getExtratoGeralCsv(
+                outputStream,
+                codUnidade,
+                dataInicial,
+                dataFinal);
     }
 
     @GET
-    @Path("/consolidados/{codUnidade}/report")
-    @Secured(permissions = Pilares.Gente.Relatorios.QUIZ)
-    public Report getExtratoGeralReport(@PathParam("codUnidade") Long codUnidade,
-                                        @QueryParam("dataInicial") long dataInicial,
-                                        @QueryParam("dataFinal") long dataFinal) {
-        return service.getExtratoGeralReport(codUnidade, dataInicial, dataFinal);
+    @Path("/extrato-geral/report")
+    public Report getExtratoGeralReport(@QueryParam("codUnidade") @Required Long codUnidade,
+                                        @QueryParam("dataInicial") @Required String dataInicial,
+                                        @QueryParam("dataFinal") @Required String dataFinal) throws ProLogException {
+        return service.getExtratoGeralReport(
+                codUnidade,
+                dataInicial,
+                dataFinal);
+    }
+
+    @GET
+    @Path("/respostas-realizados/csv")
+    @Produces("application/csv")
+    public StreamingOutput getRespostasRealizadosCsv(
+            @QueryParam("codUnidade") @Required Long codUnidade,
+            @QueryParam("codModeloQuiz") @Optional Long codModelo,
+            @QueryParam("cpfColaborador") @Optional Long cpfColaborador,
+            @QueryParam("dataInicial") @Required String dataInicial,
+            @QueryParam("dataFinal") @Required String dataFinal,
+            @QueryParam("apenasRespostasSelecionadas") @Required boolean apenasRespostasSelecionadas) {
+        return outputStream -> service.getRespostasRealizadosCsv(
+                outputStream,
+                codUnidade,
+                codModelo,
+                cpfColaborador,
+                dataInicial,
+                dataFinal,
+                apenasRespostasSelecionadas);
+    }
+
+    @GET
+    @Path("/respostas-realizados/report")
+    public Report getRespostasRealizadosReport(
+            @QueryParam("codUnidade") @Required Long codUnidade,
+            @QueryParam("codModeloQuiz") @Optional Long codModelo,
+            @QueryParam("cpfColaborador") @Optional Long cpfColaborador,
+            @QueryParam("dataInicial") @Required String dataInicial,
+            @QueryParam("dataFinal") @Required String dataFinal,
+            @QueryParam("apenasRespostasSelecionadas") @Required boolean apenasRespostasSelecionadas)
+            throws ProLogException {
+        return service.getRespostasRealizadosReport(
+                codUnidade,
+                codModelo,
+                cpfColaborador,
+                dataInicial,
+                dataFinal,
+                apenasRespostasSelecionadas);
     }
 }